pyproject.toml
[project]
name = "memory-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.5",
    "langchain-google-genai>=3.0.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]
